Improve reliability of zrm-innobackupex.
We now set wait_timeout in the socket-server to have better control.
It defaults to 8 hours, but is configurable through socket-server.conf.

We also now catch SIGPIPE/SIGTERM and cleanup gracefully, which prevents a whole
class of errors involving innobackupex continuing to run long after zrm
has gone away (and the "snapshot" subsequently not being cleaned up).

Additionally, the inno snapshot plugin will log to the socket-server log
providing additional debugging.
It also catches SIGPIPE now so that it's harder to interrupt.

Finally, the socket-server plugin no longer responds to the mysqlhotcopy
action which helps backups to fail if innobackupex didn't work for any reason.