Fix path issues, and detect init scriptness better
mysqlctl was using a bad way to auto-detect its utilities which
didn't work during startup and shutdown.

It also wasn't properly detecting when it was being called as an
init script. It worked in the superficial cases but not elsewise.

Finally, redhat/centos REQUIRE that init scripts touch a file in
/var/lock/subsys/<init name>. Tool now checks for that directory,
and touches the file if it exists and it's being called as an init script.