Fix a bug when adding partitions
When adding partitions where the last partition was older than
the current date, pdb-parted would become very confused and try to
add a partition for too far into the future.

We fix this by not attempting to add a 'catchup' partition.
The whole 'uneven' calculation has been reworked, simplified,
and better commented for future understanding.