Refactored xtrabackup-client, inno-snapshot, and xtrabackup-agent.
inno-snapshot now completely a stub. It's only purpose is to exist on
the filesystem. It will be renamed stub-snapshot.pl.

xtrabackup-client now speaks the new protocol to the agent.

Both agent and client more heavily instrumented with error checking.
Hundreds of lines of useless code deleted.

The backup model now more closely resembles "one request, one action"
models like HTTP. SIgnificant work goes into acheiving this, but the
gains are tangible, backups take less time to start, and have MANY
fewer moving pieces improving reliablility.

When doing an incremental backup and replication=0, the client will
now place a warning in it's log about the potential for unrecoverable
situations. This may be promoted to a fatal error in the future. When
an incremental is requested, and replication=1, the agent now collects
master information directly, and then restarts the slave this
eliminates needless time spent with the slave off.

Significant code re-use introduced. Both client and agent now use
ProcessLog for their logging needs. The client uses IniFile to read
it's config, and agent will follow suite.